% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicenumberrr.R
\name{to_color}
\alias{to_color}
\title{Give all parts of the number with different colors}
\usage{
to_color(number, colors = c("red", "green", "yellow", "blue"))
}
\arguments{
\item{number}{integer}

\item{colors}{vector of different colors}
}
\value{
vector of colored int
}
\description{
Give all parts of the number with different colors
}
\examples{
to_color(1234567L, c("red", "green", "yellow", "blue"))

}
