% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicenumberrr.R
\name{to_df}
\alias{to_df}
\title{Change the formatting of data in column(s) of a dataframe to either human readable or numeric}
\usage{
to_df(
  df,
  col_names = NULL,
  transform_type = "human",
  family = "number",
  errors = "raise",
  ...
)
}
\arguments{
\item{df}{dataframe, dataframe to apply formatting}

\item{col_names}{str or vector, column names to apply formatting (default is all columns)}

\item{transform_type}{str, type of transformation (e.g. human, num)}

\item{family}{str, family of suffix, numeric or filesize}

\item{errors}{'raise', 'coerce', default 'raise'
If 'raise', then invalid parsing will raise an exception.
If 'coerce', then invalid parsing will return NA.}
}
\value{
dataframe with formatting applied
}
\description{
Change the formatting of data in column(s) of a dataframe to either human readable or numeric
}
\examples{
df <- data.frame(A = c(1000, 10000), B = c(1000000, 100000))
to_df(df, col_names = c("A", "B"), transform_type = "human")
}
