% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicenumberrr.R
\name{to_human}
\alias{to_human}
\title{Convert large number to human readable string}
\usage{
to_human(n, prec = 0, family = "number", errors = "raise", custom_suff = NULL)
}
\arguments{
\item{n}{number to convert float}

\item{prec}{precision to round to}

\item{family}{family of suffix, numeric or filesize}

\item{errors}{'raise', 'coerce', default 'raise'
If 'raise', then invalid parsing will raise an exception.
If 'coerce', then invalid parsing will return NA.}

\item{custom_suff}{List of custom suffixes, default NULL}
}
\value{
string in human readable version
}
\description{
Convert large number to human readable string
}
\examples{
to_human(69420, prec = 1)
"69.4K"

}
