% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicenumberrr.R, R/to_color.R
\name{to_color}
\alias{to_color}
\title{Give all parts of the number with different colors}
\usage{
to_color(number, colors)

to_color(number, colors)
}
\arguments{
\item{number}{int}

\item{colors}{vector of different colors}
}
\value{
vector of colored int

vector of colored int
}
\description{
Give all parts of the number with different colors

Give all parts of the number with different colors
}
\examples{
to_color(1234, c('red', 'yellow'))
to_color(1234, c('red', 'yellow'))
}
