% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicenumberrr.R, R/to_df.R
\name{to_df}
\alias{to_df}
\title{Change the formatting of data in column(s) of a dataframe to either human readable or numeric}
\usage{
to_df(df, col_names, transform_type = "human", family = "numeric")

to_df(df, col_names, transform_type = "human", family = "numeric")
}
\arguments{
\item{df}{dataframe, df to apply formatting}

\item{col_names}{list or str, column names to change, default is all columns}

\item{transform_type}{str, type of transformation}

\item{family}{str, family of suffix, numeric or filesize}
}
\value{
dataframe with formatting applied

dataframe with formatting applied
}
\description{
Change the formatting of data in column(s) of a dataframe to either human readable or numeric

Change the formatting of data in column(s) of a dataframe to either human readable or numeric
}
\examples{
df <- data.frame(A = c(1000, 10000), B = c(1000000, 100000))
to_df(df, col_names=c("A", "B"), transform_type="human")


df <- data.frame(A = c(1000, 10000), B = c(1000000, 100000))
to_df(df, col_names=['A', 'B', 'C'], transform_type='human')


}
